+++ project/src/components/list-offers/list-offers.tsx
@@ -1,3 +1,4 @@
+import {memo} from 'react';
 import {Offers} from '../../types/offer';
 import Card from '../card/card';
 
@@ -19,4 +20,4 @@ function ListOffers({offers}: ListOffersProps): JSX.Element {
   );
 }
 
-export default ListOffers;
+export default memo(ListOffers, (prevProps, nextProps) => prevProps.offers === nextProps.offers);
