+++ project/src/components/page-error/page-error.tsx
@@ -1,13 +1,13 @@
 import {Fragment} from 'react';
 import {Link} from 'react-router-dom';
-import Header from '../header/header';
+import HeaderNav from '../header-nav/header-nav';
 import './style.css';
 
 
 function PageError (): JSX.Element {
   return (
     <Fragment>
-      <Header />
+      <HeaderNav />
       <section className="error container">
         <h1>404. Page not found</h1>
         <div className="error__block">
